Rename the classification of models
