+++ pydpm/example/Deep_Learning_PM/Restricted_Boltzmann_Machine/RBM_Demo.py
