+++ pydpm/example/Deep_Learning_PM/Variational_Autoencoder/VAE_Demo.py
