+++ pydpm/model/deep_learning_pm/rbm.py
