+++ pydpm/model/deep_learning_pm/vae.py
